Se agregaron dos rutas más a los usuarios | Búsqueda por ID y búsqueda por name | Se hicieron mejoras en algunas rutas

+++ api/src/controllers/userController/UserController.js
@@ -1,12 +1,64 @@
 const { User, Rol } = require("../../db")
-const { encrypt, compare } = require("../../helpers/password/bcryptHelper")
+const { encrypt, compare } = require("../../helpers/password/bcryptHelper");
+const { Op } = require("sequelize");
 
-const getAllUser = async () => {
-    const allUser = await User.findAll({
-        attributes: { exclude: ['password'] }
-    });
+const getAllUser = async (name) => {
+    try {
+        const allUser = await User.findAll({
+            attributes: { exclude: ['password'] }
+        });
+    
+        if(!allUser.length) throw new Error("Users not found");
+    
+        return allUser;
+    } catch (error) {
+        throw new Error(error.message);
+    }
+}
 
-    return allUser;
+const getOneUser = async (id) => {
+    try {
+        const user = await User.findByPk(id, {
+            attributes: { exclude: ['password'] }
+        });
+    
+        if(!user) throw new Error("User not found");
+    
+        return user; 
+    } catch (error) {
+        throw new Error(error.message);
+    }
+
+}
+
+const getSearchUser = async (name) => {
+    try {
+       const usersSearch = await User.findAll({ 
+            where: {
+                [Op.or]: [
+                    {
+                        name: {
+                            [Op.iLike]: `%${name}%`
+                        }
+                    },
+                    {
+                        surname: {
+                            [Op.iLike]: `%${name}%`
+                        }
+                    }
+                ]
+            },
+            attributes: { exclude: ['password'] } // Excluir el campo 'password' en los resultados
+        });
+
+        if(!usersSearch.length) { // Verificar si no se encontraron resultados
+            throw new Error("No users with that name or last name were found");
+        }
+
+        return usersSearch; 
+    } catch (error) {
+        throw new Error(error.message);
+    }
 }
 
 const postNewUser = async ({ name, surname, password, phone, mail, date_of_birth }) => {
@@ -28,10 +80,16 @@ const postNewUser = async ({ name, surname, password, phone, mail, date_of_birth
             }
         });
 
-
         if(!created) throw new Error("The email is already registered");
 
-        return { message: "User created successfully"};
+        return { 
+            message: "User created successfully",
+            user: {
+                name: user.name,
+                surname: user.surname  
+            }
+        } 
+
     } catch (error) {
         throw new Error(error.message);
     }
@@ -42,14 +100,19 @@ const postLoginUser = async ({ mail, password }) => {
     try {
         const user = await User.findOne({ where: { mail } });
 
-        if(!user) throw new Error("Unregistered user");
+        if(!user) throw new Error("User not registered with that email");
 
         const checkPassword = await compare(password, user.password);
 
-        if(!checkPassword) throw new Error("Password is incorrect");
+        if(!checkPassword) throw new Error("Invalid credentials");
 
-        return { message: "User successfully logged in" };
-
+        return { 
+            message: "User successfully logged in",
+            user: {
+                name: user.name,
+                surname: user.surname  
+            }
+        }       
     } catch (error) {
         throw new Error(error.message);
     }
@@ -65,7 +128,7 @@ const updateUser = async (id, userData) => {
             const checkPassword = await compare(userData.password, user.password);
 
             if(checkPassword) {
-                throw new Error("the password is the same as the current one, you must type a new one")
+                throw new Error("The password is the same as the current one, you must type a new one")
             }
 
             const passwordHash = await encrypt(userData.password);
@@ -93,6 +156,10 @@ const updateUser = async (id, userData) => {
 
 const deleteUser = async (id) => {
     try {
+        const user = await User.findByPk(id);
+
+        if(!user) throw new Error("User not found")
+
         await User.destroy({ where: { id }});
 
         return { message: "User deleted successfully"};
@@ -103,6 +170,8 @@ const deleteUser = async (id) => {
 
 module.exports = {
     getAllUser,
+    getOneUser,
+    getSearchUser,
     postNewUser,
     postLoginUser,
     updateUser,
